refactor(tests): clean up imports and reorganize package structure in test files
